+++ src/main/java/com/ctk/dao/UserRepository.java
@@ -3,12 +3,22 @@ package com.ctk.dao;
 import com.ctk.model.User;
 
 import javax.enterprise.context.SessionScoped;
+import javax.inject.Inject;
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.IOException;
 import java.io.Serializable;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 @SessionScoped
 public class UserRepository implements Serializable {
+
+    @Inject
+    private Settings settings;
 
     private List<User> users = new ArrayList<>();
 
@@ -30,5 +40,53 @@ public class UserRepository implements Serializable {
 
     public void empty() {
         users.clear();
+    }
+
+    public void loadUserFile() {
+        String line = null;
+        BufferedReader reader = null;
+
+        if (new File(String.valueOf(settings.getPathAdmin())).isFile()) {
+
+            try {
+
+                reader = Files.newBufferedReader(settings.getPathAdmin(), StandardCharsets.UTF_8);
+                line = reader.readLine();
+
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+
+            if (reader != null) {
+
+                if (line != null && !line.isEmpty()) {
+
+                    try {
+                        readingAllUsers(line, reader);
+                        reader.close();
+
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                }
+            }
+        }
+    }
+
+    private void readingAllUsers(
+            String line,
+            BufferedReader reader)
+            throws IOException {
+
+        while (line != null) {
+
+            if (!line.equals("")) {
+                List<String> tempList = Arrays.asList(line.split(";"));
+
+                add(tempList.get(0), tempList.get(1));
+            }
+            line = reader.readLine();
+        }
+
     }
 }
